fix null case input string, INT_MIN printing %i and %d

+++ main.h
@@ -47,6 +47,6 @@ int print_char(va_list arg);
 int print_percent(va_list arg);
 int print_percent_car(va_list arg);
 int print_num(va_list arg);
-char *itoa(int i, char b[]);
+char *itoa(long int i, char b[]);
 
 #endif
